Send 'user start typing' and 'user stop typing' events through web sockets
For now, the 'user start typing' event is sent when a user taps on the message text field and the 'user stor typing' event is sent when the user taps outside of the message text field. To be precise, actual events are sent when the message text field gains focus or loose it.